[feature/PPF-5]: created layout
