+++ src/app/AppRouter.tsx
@@ -4,22 +4,25 @@ import {WelcomePage} from "../pages/Welcome/WelcomePage.tsx";
 import {LoginPage} from "../pages/Login/LoginPage.tsx";
 import {RegisterPage} from "../pages/Register/RegisterPage.tsx";
 import {AuthorizedPage} from "../pages/Authorized/AuthorizedPage.tsx";
+import AuthorizedLayout from "../widgets/AuthorizedLayout/AuthorizedLayout.tsx";
 
 export const AppRouter = () => {
-  return (
-    <Routes>
-      <Route path="/" element={<WelcomePage />} />
-      <Route path="/login" element={<LoginPage />} />
-      <Route path="/register" element={<RegisterPage />} />
-      <Route 
-        path="/authorized" 
-        element={
-          <ProtectedRoute>
-            <AuthorizedPage />
-          </ProtectedRoute>
-        } 
-      />
-      <Route path="*" element={<div>404 - Страница не найдена</div>} />
-    </Routes>
-  );
+    return (
+        <Routes>
+            <Route path="/" element={<WelcomePage/>}/>
+            <Route path="/login" element={<LoginPage/>}/>
+            <Route path="/register" element={<RegisterPage/>}/>
+            <Route
+                path="/authorized"
+                element={
+                    <ProtectedRoute>
+                        <AuthorizedLayout>
+                            <AuthorizedPage/>
+                        </AuthorizedLayout>
+                    </ProtectedRoute>
+                }
+            />
+            <Route path="*" element={<div>404 - Страница не найдена</div>}/>
+        </Routes>
+    );
 };
